Changed current teal to ADA compliant Teal
There are some items on the site in a teal (#0D8896), most of which are boxes or buttons. However, this is not the same teal as the new link color- it has a contrast ratio of 4.22 with the background #FFFFFF. So, I am changing it to match the new, ADA-compliant teal of the links (#067986).